+++ app/src/components/Image/InspectScreen.tsx
@@ -3,10 +3,11 @@ import { IconNames } from "@blueprintjs/icons";
 import { useParams } from "react-router-dom";
 
 import { AppScreen, ContainerImage } from "../../Types";
-import { useStoreActions } from "../../Domain";
 import { ScreenHeader } from ".";
 import { ScreenLoader } from "../ScreenLoader";
 import { CodeEditor } from "../CodeEditor";
+
+import { useStoreActions } from "./Model";
 
 import "./InspectScreen.css";
