+++ app/src/components/Volume/ManageScreen.tsx
@@ -5,9 +5,10 @@ import { useTranslation } from "react-i18next";
 import dayjs from "dayjs";
 
 import { AppScreen } from "../../Types";
-import { useStoreActions, useStoreState } from "../../Domain";
 import { usePoller } from "../../Hooks";
 import { AppScreenHeader } from "../AppScreenHeader";
+
+import { useStoreActions, useStoreState } from "./Model";
 
 import "./ManageScreen.css";
